# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: San George Scouts API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance/sessions:
    get:
      tags: [Attendance]
      operationId: listAttendanceSessions
      summary: List all attendance sessions
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceSession'
    post:
      tags: [Attendance]
      operationId: createAttendanceSession
      summary: Create a new attendance session (leader only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceSessionInput'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceSession'
  /attendance/sessions/{sessionId}:
    get:
      tags: [Attendance]
      operationId: getAttendanceSession
      summary: Get a session with all scout attendance records
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session detail with records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceSessionDetail'
  /attendance/sessions/{sessionId}/records:
    post:
      tags: [Attendance]
      operationId: submitAttendanceRecords
      summary: Submit attendance records for a session (leader only)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceRecordsBatch'
      responses:
        '200':
          description: Records saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceSessionDetail'
...
components:
...
  schemas:
...
    AttendanceSession:
      type: object
      required: [id, title, sessionDate, createdAt, attendedCount, totalCount]
      properties:
        id:
          type: string
        title:
          type: string
        sessionDate:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        attendedCount:
          type: integer
        totalCount:
          type: integer
        excusedCount:
          type: integer
        withGearCount:
          type: integer
    AttendanceSessionInput:
      type: object
      required: [title, sessionDate]
      properties:
        title:
          type: string
          minLength: 1
        sessionDate:
          type: string
          format: date-time
        notes:
          type: string
    AttendanceRecord:
      type: object
      required: [userId, scoutName, status]
      properties:
        userId:
          type: string
        scoutName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [present, absent]
        excuse:
          type: boolean
        hasGear:
          type: boolean
    AttendanceSessionDetail:
      type: object
      required: [id, title, sessionDate, createdAt, records]
      properties:
        id:
          type: string
        title:
          type: string
        sessionDate:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        records:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceRecord'
    AttendanceRecordItem:
      type: object
      required: [userId, status]
      properties:
        userId:
          type: string
        status:
          type: string
          enum: [present, absent]
        excuse:
          type: boolean
        hasGear:
          type: boolean
    AttendanceRecordsBatch:
      type: object
      required: [records]
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceRecordItem'
...